first draft at abstracting out data loading from the main entry point, add linting
